refactor(totalDiscountPrice): 쿠폰 할인 금액을 체크 리스트에서 파생하여 selector로 가져오고 커스텀 훅을 유틸 함수로 변경
