+++ src/components/Coupon/utils/couponFinder.ts
@@ -1,0 +1,11 @@
+import { Coupon } from '@/types/coupon';
+
+export const couponFinder = (couponList: Coupon[]) => {
+  const findCouponByCode = (code: string) => {
+    return couponList.find((coupon) => coupon.code === code);
+  };
+
+  return {
+    findCouponByCode,
+  };
+};
