+++ src/recoil/coupons/selectors.ts
@@ -1,6 +1,10 @@
 import { selector } from 'recoil';
 
+import { couponSavedCheckListState } from './atoms';
 import { fetchCouponSelector } from './fetchCouponSelector';
+import { orderResultState } from '../cartItems/selectors';
+
+import discountCalculator from '@/components/Coupon/utils/couponDiscountCalculator';
 
 export const couponCheckListSelector = selector({
   key: 'couponCheckListSelector',
@@ -10,5 +14,22 @@ export const couponCheckListSelector = selector({
       ...coupon,
       isChecked: false,
     }));
+  },
+});
+
+export const totalDiscountPriceState = selector({
+  key: 'totalDiscountPriceState',
+  get: ({ get }) => {
+    const { totalOrderPrice } = get(orderResultState);
+    const couponSavedCheckList = get(couponSavedCheckListState);
+    const couponList = get(fetchCouponSelector);
+
+    const { calculateDiscountAmount } = discountCalculator(couponList);
+    return couponSavedCheckList.reduce((acc, coupon) => {
+      if (coupon.isChecked) {
+        return acc + calculateDiscountAmount(coupon, totalOrderPrice);
+      }
+      return acc;
+    }, 0);
   },
 });
